[windows] Conditionally add apm tracing to agent
Conditionally adds apm tracing to the main agent installer.
Whether component is configured is determined by the entry in
`release.json`.

When configured, adds the apm injection component via merge module,
and sets up beginning configuration.

Note:  This extends previous work to allow component to be
conditionally added.  As noted in the "comment" entry in release.json,
it is not intended to be included in release versions, and no
artifacts _will_ be included in release versions as long as the
`release.json` entries are not included.